Move file_wrapper to utils
